+++ src/AudioSettingsDialog.cpp
@@ -1,6 +1,6 @@
 /* 
  * AudioSettingsDialog.cpp is a part of LoopAuditioneer software
- * Copyright (C) 2021-2024 Lars Palo and contributors (see AUTHORS file) 
+ * Copyright (C) 2021-2025 Lars Palo and contributors (see AUTHORS file) 
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by